fix(db): Use prepared statement when inserting blocks
I did not take this for a spin for some practical testing. Seems quite
straightforwardly equivalent to me, but I wouldn't take my word for it
and hopefully unit tests will confirm.

A deadline is approaching, so I am procrastinating by poking around the
new DB code :D
And I think this could be pretty significant performance improvement for
file updates/inserts. And those seem to be a bit of a pain for users.
Though most of the pain seems to come from slow storage, and prepared
statements wont help with that (right?). And for those TrueNAS users I
suspect no amount of optimisation on our end will help, something must
be borked in ther storage setup or the layers upon virtualisation
layers. There's also users on less suspect systems that indicate that
things aren't exactly snappy, lets hope this does something for them.
And even if it has no perceptible impact, it's still nice to save a few
cpu cycles :)