info: Insert blocks in batches independently of files
tl;dr:
As predicted by Jakob, all these kinds of changes make no relevant
difference.

Some stuff that might be interesting, but mostly just a summary of the
non-outcomes that I spent (to put it neutrally) my time on:

I extended the benchmarks to also replace with a random number of blocks
and a small number of blocks. Clearly expecting that batching helps with
small number of blocks. Well the opposite is the case, benchmarking
indicates that for large number of updated blocks, batching makes no
difference or maybe a tiny improvement at the larger end of the DB.
However for replacing with just 5 blocks per file, there's no difference
at all - which to me is super surprising. I mean I didn't expect this to
make an actually significant difference in the grand scheme of things,
the main problem of random writes is there regardless of all of these
changes, but if anything should show a difference then small numbers of
blocks per file with a small (thus fast) DB. The opposite is the case,
there's even a tiny indication of no batching being faster up to a DB
size of 20k for updating with just 5 blocks.

Also mildly interesting are that there are peaks of faster performance
at certain DB sizes, and those peaks look like they were correlated
between my two benchmark runs. I ran them sequentially, so hours apart,
this wasn't a case of an external influence at a specific time
influencing both. Maybe this indicates that the db exceeds some
threshold before those peaks that triggers some sqlite housekeeping,
which then improves performance for a bit.

I ran these benchmarks on an ext4 filesystem on a mirrored LVM partition
on two (rather old) SSDs. I did so in the hopes of getting relevant
results for slow-ish filesystems, and I definitely did get the slow part
:) : About 10h per benchmark exec. I first even tried on mirrored btrfs
spinning disk, but that took ages.