fix(db): remove temp_store = MEMORY pragmas
This reduces database migration memory usage in my test scenario from
3.8 GB to 400 MB. In principle I don't think we're causing many temp
tables to be generated anyway in normal usage, but if we do and someone
can benchmark a performance difference, we can add a tunable. I ran the
database benchmark before and after and didn't see a difference above
the noise level.

Signed-off-by: Jakob Borg <jakob@kastelo.net>