chore(model): Create temp file right after space check when pulling
After a fun edge case around not creating the temp file at all was
discovered and fixed, that caused an issue with exceeding free space
limits. It's all fixed now, but nevertheless I think it would be nice to
make it a bit simpler: Create the temp file right away instead of lazily
and do the free space right before that. Writing that I kinda anticipate
that I'll trigger yet another edge-case bug with this change :)

Signed-off-by: Simon Frei <freisim93@gmail.com>