fix(sqlite): less impactful periodic garbage collection
Periodic garbage collection can take a long time on large folders. The
worst step is the one for blocks, which are typically orders of
magnitude more numerous than files or block lists.

This improves the situation in by running blocks GC in a number of
smaller range chunks, in random order, and stopping after a time limit
(compiled in to five minutes).

With this, we're not guaranteed to complete a full GC on every run, but
we'll make some progress and get there eventually.

Signed-off-by: Jakob Borg <jakob@kastelo.net>