feat(gui): show restart message during auto-upgrade (fixes #1248)
Add a new UpgradeStarted event that is emitted when an automatic upgrade
has been performed and Syncthing is about to restart. The GUI listens
for this event and shows the "Syncthing is restarting" dialog instead of
the misleading "Connection Error" message.

This improves the user experience during automatic upgrades by providing
a clear indication that the system is restarting as expected.

Signed-off-by: majiayu000 <1835304752@qq.com>